week5/src/test/java/neobis/week4/controller/OrderControllerTest.java
package neobis.week4.controller;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import static org.junit.Assert.assertEquals;

@RunWith(SpringRunner.class)
@SpringBootTest
@AutoConfigureMockMvc
 class OrderControllerTest {
    @Autowired
    OrderController orderController;

    @Autowired
    private MockMvc mockMvc;
    @Autowired
    private ObjectMapper mapper;


    @Test
    void canGetAllOrders() {
        ResponseEntity<?> responseEntity = orderController.getAllOrders();
        assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
    }

    @Test
    void canGetOrderById() {
        ResponseEntity<?> responseEntity = orderController.getOrder(2l);
        assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
    }


}
